requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-resized==1.0.2
et-xmlfile==1.1.0
idna==3.4
midtransclient==1.4.1
mysqlclient==2.1.1
Pillow==9.5.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
